+++ test/ctags/test-ctags.cpp
@@ -9,15 +9,22 @@
 #include <wex/ctags/ctags.h>
 #include <wex/ui/frd.h>
 
+#include "../syntax/test.h"
 #include "test.h"
 
 TEST_CASE("wex::ctags")
 {
+  auto* stc = new wex::test::stc();
+
+  const wex::path p("test.h");
+  ALLOW_CALL(*stc, path()).RETURN(p);
+
+  wex::ctags_entry::register_image(stc);
+
   SUBCASE("constructor")
   {
     // default auto_complete
-    REQUIRE(
-      wex::ctags(get_stc()).auto_complete("test_").starts_with("test_app"));
+    REQUIRE(wex::ctags(stc).auto_complete("test_").starts_with("test_app"));
 
     // setup a filter using find
     wex::ctags_entry filter;
@@ -26,11 +33,10 @@ TEST_CASE("wex::ctags")
     REQUIRE(filter.kind() == "f");
 
     // auto_complete using filter should now return member functions
-    REQUIRE(
-      wex::ctags(get_stc()).auto_complete("me", filter).starts_with("method"));
-    REQUIRE(wex::ctags(get_stc()).auto_complete("he", filter).empty());
+    REQUIRE(wex::ctags(stc).auto_complete("me", filter).starts_with("method"));
+    REQUIRE(wex::ctags(stc).auto_complete("he", filter).empty());
 
-    REQUIRE(wex::ctags(get_stc()).separator() != ' ');
+    REQUIRE(wex::ctags(stc).separator() != ' ');
     REQUIRE(wex::ctags::close());
   }
 
@@ -47,7 +53,7 @@ TEST_CASE("wex::ctags")
     REQUIRE(!wex::ctags::previous());
     REQUIRE(!wex::ctags::find("xxxx"));
     REQUIRE(wex::ctags::find("test_app"));
-    REQUIRE(wex::ctags::find("test_app", get_stc()));
+    REQUIRE(wex::ctags::find("test_app", stc));
     REQUIRE(!wex::ctags::next());
     REQUIRE(!wex::ctags::previous()); // there is only 1 match?